refactor: constants 폴더 이름 변경
